src/requirements.txt
astroid==2.1.0
autopep8==1.4.3
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.2
django-allauth==0.38.0
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-tinymce4-lite==1.7.4
idna==2.8
isort==4.3.4
jsmin==2.2.2
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==3.0.0
pep8==1.7.1
Pillow==5.4.1
pycodestyle==2.4.0
pylint==2.2.2
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.5
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.2.4
typed-ast==1.2.0
urllib3==1.24.1
wrapt==1.11.1
